Fix to #15778 - Query: null semantics visitor incorrectly classifies subquery returning single scalar value as non-nullable
Problem was that when we compared scalar subquery to null, we would assume the subquery is never null so we would never return any results. Scalar subqueries can be null when using entities.Where(e => false).Select(e => e.Name).FirstOrDefault() pattern.

Fix is to always assume SubSelectExpression is nullable.
Also fixed a bug in subquery member pushdown that was exposed by the null semantics fix - because now subqueries are treated as nullable, null semantics applies on them and in some cases of collection pushdown we would match null == null as true and returning results where nothing should be returned.
Fix is to add null check for one of the keys, so that we only match outer and inner when they are not null - this is for case of navigation chaining so its always correct to do it. We never expect a child entity when it's parent is null, or where the FK to that entity is null